Refactor KMEA Keyboard JS handlers
Refactor common JS interface functions from `KMInAppKeyboardJSHandler` and `KMSystemKeyboardJSHandler` into a base class `KMKeyboardJSHandler`.

They still need to implement `setIsChiral` and `insertText`.
